% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_all_random}
\alias{merge_all_random}
\title{Merge all random}
\usage{
merge_all_random(pids, pids_dad, birthyears, paternalped_ids,
  surr_pid_start = 50000000L, max_it = -1L)
}
\description{
Merge all random
}
