% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_certain_pedigree_randomly}
\alias{merge_certain_pedigree_randomly}
\title{Merge certain pedigree randomly}
\usage{
merge_certain_pedigree_randomly(pids, pids_dad, birthyears,
  paternalped_ids, pedid_to_merge, no_surrogate_ancestors = 1L,
  surr_pid_start = 50000000L, verbose = FALSE)
}
\arguments{
\item{pids}{vector of person ids: >= 1 (`NA` not allowed)}

\item{pids_dad}{vector of person ids for dad: >= 1 (`NA` for founders)}

\item{birthyears}{vector of birth years: free}

\item{paternalped_ids}{vector of pedigree ids: >= 1 (`NA` not allowed)}

\item{pedid_to_merge}{id of pedigree that will be merged with other pedigrees}

\item{no_surrogate_ancestors}{number of surrogate ancestors to add}

\item{surr_pid_start}{start person id for new surrogate individuals}

\item{verbose}{verbose output}
}
\value{
Only individuals in new pedigree.
}
\description{
Merge certain pedigree randomly
}
