% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_pedigree}
\alias{merge_pedigree}
\title{Merge a pedigree randomly}
\usage{
merge_pedigree(pids, pids_dad, birthyears, paternalped_ids, pedid_to_merge,
  sons_configs, no_surrogate_ancestors = 1L,
  surr_pid_start = 50000000L, verbose = FALSE)
}
\arguments{
\item{pids}{vector of person ids: >= 1 (`NA` not allowed)}

\item{pids_dad}{vector of person ids for dad: >= 1 (`NA` for founders)}

\item{birthyears}{vector of birth years: free}

\item{paternalped_ids}{vector of pedigree ids: >= 1 (`NA` not allowed)}

\item{pedid_to_merge}{id of pedigree that will be merged with other pedigrees}

\item{sons_configs}{list of sons (age of father when each son was born); 
elements cannot be of size 0 and all ages be > 0. E.g. `list(c(28, 30), c(24))`.}

\item{no_surrogate_ancestors}{number of surrogate ancestors to add}

\item{surr_pid_start}{start person id for new surrogate individuals}

\item{verbose}{verbose output}
}
\value{
Only individuals in new pedigree.
}
\description{
Note that children will be added according to `sons_configs`.
}
