% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_single_random}
\alias{merge_single_random}
\title{Merge single random}
\usage{
merge_single_random(pids, pids_dad, birthyears, paternalped_ids,
  pid_to_merge, max_meioses = 100L, surr_pid_start = 50000000L)
}
\description{
Merge single random
}
