% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proto_merge_all_random}
\alias{proto_merge_all_random}
\title{Merge all random}
\usage{
proto_merge_all_random(pids, pids_dad, birthyears, paternalped_ids,
  max_it = -1L, surr_pid_start = 50000000L, verbose = FALSE)
}
\arguments{
\item{pids}{vector of person ids: >= 1 (`NA` not allowed)}

\item{pids_dad}{vector of person ids for dad: >= 1 (`NA` for founders)}

\item{birthyears}{vector of birth years: free}

\item{paternalped_ids}{vector of pedigree ids: >= 1 (`NA` not allowed)}

\item{max_it}{maximum number of iterations (merge events), `-1` for disable}

\item{surr_pid_start}{start person id for new surrogate individuals}

\item{verbose}{verbose output}
}
\value{
All individuals, including new surrogate.
}
\description{
Merge all random
}
