% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_index}
\alias{random_index}
\title{Random index between 0 and n-1 (both included)}
\usage{
random_index(n)
}
\arguments{
\item{n}{upper limit (not included)}
}
\description{
Random index between 0 and n-1 (both included)
}
