% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_merge_input}
\alias{validate_merge_input}
\title{Validate merge input}
\usage{
validate_merge_input(pids, pids_dad, birthyears, paternalped_ids)
}
\arguments{
\item{pids}{vector of person ids: >= 1 (`NA` not allowed)}

\item{pids_dad}{vector of person ids for dad: >= 1 (`NA` for founders)}

\item{birthyears}{vector of birth years: free}

\item{paternalped_ids}{vector of pedigree ids: >= 1 (`NA` not allowed)}
}
\description{
Validate merge input
}
