% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_sons_configs}
\alias{validate_sons_configs}
\title{Validate configuration of sons (age of father when each son is born)}
\usage{
validate_sons_configs(sons_configs)
}
\arguments{
\item{sons_configs}{list of sons (age of father when each son was born); 
elements cannot be of size 0 and all ages be > 0. E.g. `list(c(28, 30), c(24))`.}
}
\description{
Validate configuration of sons (age of father when each son is born)
}
