% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_surr_pid_start}
\alias{validate_surr_pid_start}
\title{Validate surrogate father id input}
\usage{
validate_surr_pid_start(pids, surr_pid_start)
}
\arguments{
\item{pids}{person id: >= 1 (`NA` not allowed)}

\item{surr_pid_start}{start person id for new surrogate individuals}
}
\description{
Validate surrogate father id input
}
